change consts to storage, change winnerFeeSahere from uint16 to uint23, onlyOwner actions, most params are now in 1e6 format

+++ src/interfaces/IArbiterAmAmmHarbergerLease.sol
@@ -37,7 +37,7 @@ interface IArbiterAmAmmHarbergerLease {
     /// @return The percentage of the swap fee that will be paid to the winner
     function winnerFeeShare(
         PoolKey calldata key
-    ) external view returns (uint16);
+    ) external view returns (uint24);
 
     /// @return The deposit of the account for the asset
     /// @param asset The address of the ERC20 to check
